+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.9.2 2024-06-12
+
+# New features
+- #121 auto upper/lower for text fields
+
+## Fixed issues
+
+- #118 date format RR not working for APEX<21.2
+- #119 delete button on array items does not work for newly created 
+
 # 0.9.1 2024-06-03
 
 # New features